Finalize command panel option
I want to be able to wrap them such that we can add UI handling code to
the actions with the UI layer without requiring the rest of the code
that creates choices for the player to be aware of it.  To use a var in
a lambda, it needs to be final.  Since I don't want these things to be
modified anyway (they are meant to be created on demand), require
everything to be final.