BUGFIX: Add awful hack to fix getting stuck watching combat.
Players can observe combat scenes through the observer interface.
Essentially the only way to check this is through the beingObserved
flag, which is not a great pattern.  A previous commit changed this from
a Global.gui() method to a method on the combatants.  However, if the
player is not a combatant, they simply never get the appropriate prompt
and get stuck in the combat.