Rework command panel options to a scrolling style.
This dedicates much more space for the buttons themselves and moves
toward the design specified.  The buttons automatically resize to fit
the size of the window and add scrollbars when necessary, which is an
improvement.  There's also substantially less code, so this feels more
natural within the swing framework.

In general, this design wastes less screen real estate while also being
more adaptive to different window sizes.