Get rid of skill tooltips
Tool tips are a non-starter for keyboard only control. It's highly
desirable to orient this game toward one-handed play. For most people,
their dominant hand is occupied, which is also the one usually
controlling the mouse.

This is still a bit janky at the moment because of the way the text
wrapping works (or doesn't in this case) but I maintain that it's
better. I may need to rework the GUI design around a "detail view" but
I want to think about it more.