Center rather than taking up all the space.
While it was nice having big ol' buttons to click, they looked a little
weird with so much empty space.  This change causes the buttons to use
their preferred sizes.  However, a consequence of how the default
viewport layout worked would put these all the way on the right of the
screen (the viewport was assuming its preferred size, the whole bottom
panel) and then putting the contents in the top left. We want the
contents centered, and so require an extension of the viewport layout to
 implement that.