Look up proper factory for SizeMods
Since there's no default constructor, all this reflection stuff failed
when the default constructor for size mods was removed.  Since there are
 now requirements for creating a SizeMod, call out to a static factory
 method that can see a private zero arg constructor.

Really, the right way to do this is to have a constructor with a
JsonElement input and look for that one, but I'll deal with that at
another time.