Copy body parts when copying the body
It's come to my attention that combat makes a copy of the combatants
body, but that said copy does not copy the body parts.  This relies on
the copy on write behavior of applying mods and such to evaluate the
NPCs actions.  Get around that via an explicit copy for evaluation
purposes.

This only copies mutable body parts, i.e. those that derive from
GenericBodyPart.  It will suffice for now.