Introduce ambiguity to the various stat meter bars
The characters don't always have perfect information about each other.
The AI can represent this as number ranges, but players are going to
need a visual clue.  A flashing section works well as an indicator that
it might be full, or it might be empty.  The solid section represents
what the character knows to be true.