Remove the focus system
Choices like this are weird skills.  It's a little bit like a wager,
which I think fits into the game, but the wager action taking a turn
that would be otherwise used to damage the opponent is such a
disadvantage that I'm not sure why anyone would use a turn on this.

It also breaks the flow of battle a little bit.  You just stop to get
beat on for a round so you can _think_ about what you want to do in the
future? I'll likely eventually build a wagering mechanic, but it won't
look like this.