Get rid of helpers that aren't pulling their weight
This change relies on the ability of Sizes to self-regulate to elide
some size checks.  Beyond that, it just inlines the check and assumes
that random is often as good as iterating through the list.  There's
very few cases where a character has multiple of the same part and we'll
 come up with a better solution for that.