Replace BodyReplaceEffects with BodyGrowthEffects
We don't really have a situation where multiple of the same part are
allowed, so we're never going to get into a state where we're actually
replacing something meaningful.

One side effect of this change is that your existing mods will carry
over when you get transformed.  I think this is a good thing, but we'll
see the effects when I can progress that far in the game.