Remove genetic limits on physical and sexual stamina
In the imaginary world where we have sufficient technology and magic to
balloon a person's tits in a matter of a combat round, we somehow still
have innate limits on physical and sexual stamina that vary by
individual?  Give me a break.

Also, I don't want to maintain a link between a character's natural
development and their capacity to do experience gainz from deadlifts. It
breaks the encapsulation I want.