Remove a ton of code to override hard-coded bodies.
It's not that making all this stuff configurable isn't a good feature,
but setting up an entire separate system for body config from the one
that already exists is not the way to do this.  We will tackle this
problem when we've made sense of the character description.  Until then,
 simplicity is the name of the game.