Extract Cute pain effects from Character
This also integrates a new dependency: SimpleNLG.  SimpleNLG is a
natural language generation library that will handle things like
conjugating verbs and handling agreement, etc. In this case, it's used
to conjugate use, which is a relatively simple case, but we can stop
maintaining our own lexicon by using theirs.