Convert the CommandPanel to JavaFX.
It's the future! For one, it already feels snappier. Also, it was
a cinch to set up the keyboard bindings and make that navigation tight.
This will not only make the game faster to debug as I can breeze through
the menus faster, but it will make the play experience nicer (and more
one-handed, which I know is important to our players ;) )