Terminate the program when the window closes
This works around what I think is an issue with XAWT in JDK 13, but I'm
really not sure. I'm okay with this hack because I want to phase out AWT
and Swing in favor of JFX. If it causes problems, there is probably a
reason the XAWT thread is staying open and we can figure out how to
close it.