Move combat creation from the Trap to the Encounter
I've decided that encounters should create combat instances. There's
already enough of them working this way, and it seems a reasonable
thing to have happen.  Therefore, other things shouldn't create them.

To that end, I've relocated the combat creation in traps to be in the
encounter and have the trap pass back whatever information it needs
to set up the combat.