Flip mercy (in some cases)
Mercy seems to be used in two ways. Some operations add the victor in
combat, others add the loser.  This is certainly not how it should
work, but probably happened because the name is vague.  invalidAttackers
is much clearer about what sort of data it contains, so let's rename
it to that, fix the type up properly (this is never used concurrently
and copy() generates a new one), and try to make it make sense.