Bounty uses delegation instead of Globals
By introducing a hunter class, we can control the number of points
given out for a win.  We can use this to eliminate the match lookup:
hunters and prey give out different numbers of points that normal
Participants, and we can even give the prey it's +1 bonus for a win.