Don't use invalidAttackers for other purposes
We already have all the information we need to calculate a bounty and
award points when a Participant is defeated. Doling out the points then:

a) Gives a better signal to the player about what they get points for
b) Fixes a bug where players get the wrong number of points when they
resupply and clear their invalid attackers
c) Simplifies the concerns around invalid attackers: now it only has to
worry about who is invalid, not who should get points.