"handle" Enthralled
I'm so sick of handleEnthralled.  It's forcing my hand stupidly in
how Participants interact with their Intelligence, and I finally though
up a good way for dealing with it, so _deal_ with it.

The good news is, I think this will come in handy when it comes to Trap
putting Participants in Statuses and filtering their moveset.