Clean up what looked like a hack job.
These are the only two statuses that handle do something on a null, and
only a handful more that attempt to check for the presence of a null.
Looks like someone was trying to hastily jam in some feature.  There's
no rush on this, so let's do it right and split it into two methods.

Because I don't need to deal with the rest of combat quite yet, I'm
leaving the tick(null) in, so that the rest of the character ticking
happens on Match rounds.