Remove debug movement
I've looked it over, and decided that this isn't necessary to maintain.

While hooked up to a debugger, I can use the expression evaluation
tool to set any variable to basically anything I want.  Therefore, this
debug GUI is simply a set of convenience functions.  Having a convenience
GUI is completely unnecessary, as it just provides for text entry like
the evaluate expression tool.  I definitely don't need to learn and
remember another command syntax.

When I decide I need this convenience, I'll write it in a way that
utilizes the debugger's GUI, but for now, I'm just going to remove
things because YAGNI