Use overloads instead of overrides
After I think about it enough, I realize I can't make any of the
guarantees that I want to make about the overrides.  For example, they
aren't commutative, which is weird, because logical "and" and "or"
_are_ commutative.

I think the guarantees around things like commutativity are more
important that maybe getting some more diagnostic information after
a Predicate has been down-casted. I feel like it's very unlikely we'll
be intentionally stringing it at that point, as it's likely going to
library code.