Decider.parseMoves(): Remove duplicate condition
One of these conditions searches for the base, the other one searches
for the resupply _at_ the base. However, really what we're looking for
is the resupply action, or an action to lead us to the resupply action,
and the bestMoves() function is totally capable of doing that.  The
resupply actions usable() function stops us from finding someone else's
base.

It also reads nicely: if you're a hunter and have the flag, go to
a Resupply