Simulacrum: Remove skill
I don't think this skill has a place in the game right now. For one,
it's presence is making reworking pets more difficult, which is what
brought me to have to make a choice about it.

From there, I had to look into the purpose it served in the game, and
I think it fails at least two marks there. I'm not sure how many
conditions I have, but there can't be that many and failing two is
certainly not good.

First, I'm not sure it fits with Angel thematically.  One of her late
game lines is summon-focused already, and I think it's a sign of
confused character design when the same character has too-closely
related mechanics.

Second, I think it's really overpowered, and Angel is already crazy
strong.  Summoning a full-strength copy of the opponent to fight against
the opponent _and_ Angel is going to be really difficult to overcome.
Generally, one skill shouldn't roughly double the strength of one's
team without having either some risk or some long term cost.

So yeah, 1) not thematic, 2) overpowered, and 3) programmatically
complicated is three strikes and that's enough for me to kill it.